refactor(ingress): move the Traefik route writer into its own module
`benchpress/ingress.py` owns every write to the Traefik route directory. The
route files, the wildcard certificate anchor, the atomic write and the TLS
probe leave `deploy_manager`, which keeps only the job functions that a dotted
string names.

Four names lose their underscore, because `deploy_manager` now calls them
across a module boundary: `_write_instance_route` to `publish`,
`_delete_instance_route` to `withdraw`, `_ensure_wildcard_anchor` to
`ensure_anchor`, and `_log_certificate_state` to `log_certificate_state`.

`published()` and `protected_present()` are new. They let
`reconcile_instance_routes` stop globbing the directory, so the directory path
and `PROTECTED_ROUTE_FILES` appear in one file only. `publish` reads
`addressing.IDE_HTTP_PORT` instead of a hardcoded 8080.

No dotted string moves. The four job strings still name
`benchpress.deploy_manager.*`, and the three RQ job ids are unchanged. There is
no re-export shim: callers reach the new module as `ingress.publish(...)`.
