+++ benchpress/tests/test_admission.py
@@ -555,7 +555,7 @@ class TestAdmission(IntegrationTestCase):
 			patch("benchpress.deploy_manager.stop_container"),
 			patch("benchpress.deploy_manager.remove_container"),
 			patch("benchpress.deploy_manager._drop_site_database"),
-			patch("benchpress.deploy_manager._delete_instance_route"),
+			patch("benchpress.ingress.withdraw"),
 			patch("frappe.db.commit"),
 		):
 			teardown_bench(frappe.get_doc(BENCH, bench.name), **kwargs)
